refactor(web): simplify utility naming — single file is utils.ts, not utils- prefix
